wh-slot-booking/backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
passlib[bcrypt] 
python-jose[cryptography] 
python-multipart 
pydantic 
alembic